+++ CHANGELOG.md
@@ -12,11 +12,12 @@
 
 -->
 
-## Not released yet
+## v1.13.4 (20-Nov-2025)
+
 ### Changed
 - updates to `bit-get-cov-stats`
   - can start from bam file now in addition to mosdepth per-base.bed.gz (will generate the mosdepth output if starting from bam)
-  - modularized
+  - modularized, integration test added
 - updates to `bit-check-for-fastq-dup-headers`
   - autodetect gzipped or not
   - modularized, test added